+++ CHANGELOG.md
@@ -13,6 +13,8 @@
   ([rhbz#2097733])
 - Set HTTP headers to prevent sending referrer ([rhbz#2097732])
 - Set cookie option SameSite to Lax ([rhbz#2097730])
+- Set `Content-Security-Policy: frame-ancestors 'self'; default-src 'self'`
+  HTTP header for all responses ([rhbz#2097778])
 
 ### Added
 - Add support for fence\_mpath to `pcs stonith update-scsi-devices` command
@@ -50,6 +52,7 @@
 [rhbz#2097731]: https://bugzilla.redhat.com/show_bug.cgi?id=2097731
 [rhbz#2097732]: https://bugzilla.redhat.com/show_bug.cgi?id=2097732
 [rhbz#2097733]: https://bugzilla.redhat.com/show_bug.cgi?id=2097733
+[rhbz#2097778]: https://bugzilla.redhat.com/show_bug.cgi?id=2097778
 [huntr#220307]: https://huntr.dev/bounties/7aa921fc-a568-4fd8-96f4-7cd826246aa5/